Moved channel chooser to prompt class
